Added capability to recognise having played video Added capability for updating stats after playing video Began implementation of option to use cached version of video
Fixed issue where downloading one audio or video wrongly showed all
files as downloaded. This was due to the check in download_bloc being
based on voice name and media file length, which was the same for all
items. Changed it to be a comparison of ids, which should be more
reliably different for each MediaItem.